🥅 Catch transaction fail event

+++ components/Gnb/Account/PendingTx/TxItem.tsx
@@ -1,11 +1,10 @@
 /* eslint-disable react/jsx-no-target-blank */
-import { MouseEvent, useState } from 'react'
-import { useMount } from 'react-use'
+import { MouseEvent } from 'react'
 import { formatDistanceToNow } from 'date-fns'
 import styles from './style.module.scss'
 
-import { parseTxError } from 'utils/error'
 import { txInfoMessage, txToastTitle } from 'utils/transaction'
+import { parseTxError } from 'utils/error'
 import { getTxUrl } from 'utils/url'
 import { useTx } from 'hooks'
 import { renderToastEmoji } from 'components/Toast/utils'
@@ -18,61 +17,26 @@ type TxItemProps = {
 
 export function TxItem({ transaction }: TxItemProps) {
   const { txService } = useTx()
-  const [error, setError] = useState<any>(null)
 
-  const {
-    action,
-    addedTime,
-    finalizedTime,
-    hash,
-    params,
-    status,
-    error: storedError,
-  } = transaction
+  const { action, addedTime, finalizedTime, hash, status, params, error } =
+    transaction
   const txUrl = getTxUrl(hash)
-  const timestamp = finalizedTime || addedTime
-
-  const txType = error ? 'error' : 'info'
+  const isFailed = status === 'error'
+  const txType = isFailed ? 'error' : 'info'
   const title = txToastTitle(action, txType)
-  const message = error
+  const message = isFailed
     ? parseTxError(error)!.message
     : txInfoMessage(action, params)
-
-  async function checkTxStatus() {
-    if (!txService) return
-    if (finalizedTime && status === 'error' && !!storedError) {
-      setError(storedError)
-      return
-    }
-
-    try {
-      await txService.getTxReceipt(hash)
-    } catch (err) {
-      // NOTE: Failed tx
-      setError(true)
-      const newTx: Transaction = {
-        ...transaction,
-        status: 'error',
-        finalizedTime: Date.now(),
-        error: err,
-      }
-
-      const hashKey = txService.encodeKey(hash, action)
-      txService.updateTx(hashKey, newTx)
-    }
-  }
+  const timestamp = finalizedTime || addedTime
 
   function handleDelete(e: MouseEvent<HTMLButtonElement>) {
     e.stopPropagation()
 
     if (!txService) return
-    const hashKey = txService.encodeKey(hash, action)
+    const hashKey = txService.findHashKey(hash)
+    if (!hashKey) return
     txService.removeTx(hashKey)
   }
-
-  useMount(() => {
-    checkTxStatus()
-  })
 
   return (
     <li className={styles.txItem}>
